+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## Unreleased
+
+#### New Features
+
+#### Improvements
+- Added connection (lost) indicator on the Medusa log ([10774](https://github.com/pymedusa/Medusa/pull/10774))
+- Extend subtitle file parsing to allow for titles with language name. ([10782](https://github.com/pymedusa/Medusa/pull/10782))
+
+#### Fixes
+- Homepage: Fix loading shows from localCache ([10779](https://github.com/pymedusa/Medusa/pull/10779))
+- Fix Erai-raws formatted anime release guessit parsing ([10791](https://github.com/pymedusa/Medusa/pull/10791))
+ 
+-----
+
 ## 1.0.3 (10-06-2022)
 
 #### Improvements